Move filters from dispatcher to core module
