+++ data-plane/core/pom.xml
@@ -148,6 +148,10 @@
 
     <dependency>
       <groupId>io.cloudevents</groupId>
+      <artifactId>cloudevents-sql</artifactId>
+    </dependency>
+    <dependency>
+      <groupId>io.cloudevents</groupId>
       <artifactId>cloudevents-kafka</artifactId>
       <exclusions>
         <exclusion>
