+++ data-plane/core/src/main/java/dev/knative/eventing/kafka/broker/core/filter/Filter.java
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package dev.knative.eventing.kafka.broker.dispatcher;
+package dev.knative.eventing.kafka.broker.core.filter;
 
 import io.cloudevents.CloudEvent;
 import java.util.function.Predicate;
