+++ data-plane/core/src/test/java/dev/knative/eventing/kafka/broker/core/filter/subscriptionsapi/AllFilterTest.java
@@ -13,22 +13,27 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package dev.knative.eventing.kafka.broker.dispatcher.impl.filter.subscriptionsapi;
+package dev.knative.eventing.kafka.broker.core.filter.subscriptionsapi;
 
 import static org.assertj.core.api.Assertions.assertThat;
 
+import dev.knative.eventing.kafka.broker.core.filter.subscriptionsapi.AllFilter;
+import dev.knative.eventing.kafka.broker.core.filter.subscriptionsapi.ExactFilter;
+import dev.knative.eventing.kafka.broker.core.filter.Filter;
 import io.cloudevents.CloudEvent;
 import io.cloudevents.core.builder.CloudEventBuilder;
 import java.net.URI;
 import java.time.OffsetDateTime;
 import java.time.ZoneOffset;
+import java.util.Collections;
+import java.util.List;
 import java.util.Map;
 import java.util.stream.Stream;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.Arguments;
 import org.junit.jupiter.params.provider.MethodSource;
 
-public class PrefixFilterTest {
+public class AllFilterTest {
 
     static final CloudEvent event = CloudEventBuilder.v1()
             .withId("123-42")
@@ -42,16 +47,31 @@ public class PrefixFilterTest {
 
     @ParameterizedTest
     @MethodSource(value = {"testCases"})
-    public void match(CloudEvent event, String key, String value, boolean shouldMatch) {
-        var filter = new PrefixFilter(Map.of(key, value));
+    public void match(CloudEvent event, Filter filter, boolean shouldMatch) {
         assertThat(filter.test(event)).isEqualTo(shouldMatch);
     }
 
     static Stream<Arguments> testCases() {
         return Stream.of(
-                Arguments.of(event, "id", "123", true),
-                Arguments.of(event, "id", "124", false),
-                Arguments.of(event, "source", "/api", true),
-                Arguments.of(event, "source", "/news", false));
+                Arguments.of(event, new AllFilter(List.of(new ExactFilter(Map.of("id", "123-42")))), true),
+                Arguments.of(
+                        event,
+                        new AllFilter(List.of(
+                                new ExactFilter(Map.of("id", "123-42")),
+                                new ExactFilter(Map.of("source", "/api/some-source")))),
+                        true),
+                Arguments.of(
+                        event,
+                        new AllFilter(List.of(
+                                new ExactFilter(Map.of("id", "123")),
+                                new ExactFilter(Map.of("source", "/api/some-source")))),
+                        false),
+                Arguments.of(
+                        event,
+                        new AllFilter(List.of(
+                                new ExactFilter(Map.of("id", "123-42")),
+                                new ExactFilter(Map.of("source", "/api/something-else")))),
+                        false),
+                Arguments.of(event, new AllFilter(Collections.emptyList()), true));
     }
 }
