+++ data-plane/core/src/test/java/dev/knative/eventing/kafka/broker/core/filter/subscriptionsapi/PrefixFilterTest.java
@@ -13,10 +13,11 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package dev.knative.eventing.kafka.broker.dispatcher.impl.filter.subscriptionsapi;
+package dev.knative.eventing.kafka.broker.core.filter.subscriptionsapi;
 
 import static org.assertj.core.api.Assertions.assertThat;
 
+import dev.knative.eventing.kafka.broker.core.filter.subscriptionsapi.PrefixFilter;
 import io.cloudevents.CloudEvent;
 import io.cloudevents.core.builder.CloudEventBuilder;
 import java.net.URI;
@@ -28,7 +29,7 @@ import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.Arguments;
 import org.junit.jupiter.params.provider.MethodSource;
 
-public class SuffixFilterTest {
+public class PrefixFilterTest {
 
     static final CloudEvent event = CloudEventBuilder.v1()
             .withId("123-42")
@@ -43,15 +44,15 @@ public class SuffixFilterTest {
     @ParameterizedTest
     @MethodSource(value = {"testCases"})
     public void match(CloudEvent event, String key, String value, boolean shouldMatch) {
-        var filter = new SuffixFilter(Map.of(key, value));
+        var filter = new PrefixFilter(Map.of(key, value));
         assertThat(filter.test(event)).isEqualTo(shouldMatch);
     }
 
     static Stream<Arguments> testCases() {
         return Stream.of(
-                Arguments.of(event, "id", "42", true),
-                Arguments.of(event, "id", "43", false),
-                Arguments.of(event, "source", "source", true),
-                Arguments.of(event, "source", "sink", false));
+                Arguments.of(event, "id", "123", true),
+                Arguments.of(event, "id", "124", false),
+                Arguments.of(event, "source", "/api", true),
+                Arguments.of(event, "source", "/news", false));
     }
 }
