+++ data-plane/core/src/test/java/dev/knative/eventing/kafka/broker/core/filter/subscriptionsapi/SuffixFilterTest.java
@@ -13,25 +13,23 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package dev.knative.eventing.kafka.broker.dispatcher.impl.filter.subscriptionsapi;
+package dev.knative.eventing.kafka.broker.core.filter.subscriptionsapi;
 
 import static org.assertj.core.api.Assertions.assertThat;
 
-import dev.knative.eventing.kafka.broker.dispatcher.Filter;
+import dev.knative.eventing.kafka.broker.core.filter.subscriptionsapi.SuffixFilter;
 import io.cloudevents.CloudEvent;
 import io.cloudevents.core.builder.CloudEventBuilder;
 import java.net.URI;
 import java.time.OffsetDateTime;
 import java.time.ZoneOffset;
-import java.util.Collections;
-import java.util.List;
 import java.util.Map;
 import java.util.stream.Stream;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.Arguments;
 import org.junit.jupiter.params.provider.MethodSource;
 
-public class AnyFilterTest {
+public class SuffixFilterTest {
 
     static final CloudEvent event = CloudEventBuilder.v1()
             .withId("123-42")
@@ -45,37 +43,16 @@ public class AnyFilterTest {
 
     @ParameterizedTest
     @MethodSource(value = {"testCases"})
-    public void match(CloudEvent event, Filter filter, boolean shouldMatch) {
+    public void match(CloudEvent event, String key, String value, boolean shouldMatch) {
+        var filter = new SuffixFilter(Map.of(key, value));
         assertThat(filter.test(event)).isEqualTo(shouldMatch);
     }
 
     static Stream<Arguments> testCases() {
         return Stream.of(
-                Arguments.of(event, new AnyFilter(List.of(new ExactFilter(Map.of("id", "123-42")))), true),
-                Arguments.of(
-                        event,
-                        new AnyFilter(List.of(
-                                new ExactFilter(Map.of("id", "123-42")),
-                                new ExactFilter(Map.of("source", "/api/some-source")))),
-                        true),
-                Arguments.of(
-                        event,
-                        new AnyFilter(List.of(
-                                new ExactFilter(Map.of("id", "123")),
-                                new ExactFilter(Map.of("source", "/api/some-source")))),
-                        true),
-                Arguments.of(
-                        event,
-                        new AnyFilter(List.of(
-                                new ExactFilter(Map.of("id", "123-42")),
-                                new ExactFilter(Map.of("source", "/api/something-else")))),
-                        true),
-                Arguments.of(
-                        event,
-                        new AnyFilter(List.of(
-                                new ExactFilter(Map.of("id", "123")),
-                                new ExactFilter(Map.of("source", "/api/something-else")))),
-                        false),
-                Arguments.of(event, new AnyFilter(Collections.emptyList()), false));
+                Arguments.of(event, "id", "42", true),
+                Arguments.of(event, "id", "43", false),
+                Arguments.of(event, "source", "source", true),
+                Arguments.of(event, "source", "sink", false));
     }
 }
